feat: 모든 미팅방 확인하기 api

+++ src/main/kotlin/com/gcc/miti/module/dto/makegroupdto/GroupDto.kt
@@ -8,7 +8,7 @@ import java.time.LocalDateTime
 data class GroupDto(
     val description: String,
     val title: String,
-    val maxUsers: Long,
+    val maxUsers: Int,
     val meetDate: LocalDateTime?,
     val meetPlace: String?,
 ) {
